template

+++ resources/views/eskul.blade.php
@@ -1,26 +1,53 @@
-<h1>Data Eskul</h1>
-@if (session('status'))
-    <div class="alert alert-success">
+@extends('layouts/main')
+
+@section('tittle','Eskul')
+
+@section('page-tittle','Data Eskul')
+
+@section('breadcrumb')
+<li class="breadcrumb-item"><a href="#">Kategori Penilaian</a></li>
+<li class="breadcrumb-item active" aria-current="page">Eskul</li>
+@endsection
+
+
+@section('content')
+@if(session('status'))
+    <div class="alert alert-success alert-dismissible fade show" role="alert">
         {{ session('status') }}
+        <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
     </div>
 @endif
-<a href="{{ url('eskul/add') }}">Tambah</a>
-<table border="1">
-    <thead>
-        <td>#</td>
-        <td>Nama Eskul</td>
-        <td>Action</td>
-    </thead>
-    <tbody>
-        @foreach($eskul as $data)
-        <tr>
-            <td>{{$loop->iteration}}</td>
-            <td> {{ $data->nama }}</td>
-            <td>
-                <a href="eskul/edit/{{$data->id}}">Edit</a> ||
-                <a href="eskul/hapus/{{$data->id}}">Hapus</a>
-            </td>
-        </tr>
-        @endforeach
-    </tbody>
-</table>
+<div class="row">
+    <!-- column -->
+    <div class="col-sm-12">
+        <div class="card">
+            <div class="card-body">
+                <a href="{{ url('eskul/add') }}" class="btn btn-primary mb-4"><i class="mdi mdi-plus"></i> Tambah</a>
+                <div class="table-responsive">
+                    <table class="table user-table" id="table-datatables">
+                        <thead>
+                            <tr>
+                                <th>#</th>
+                                <th>Nama Eskul</th>
+                                <th>Action</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            @foreach($eskul as $data)
+                            <tr>
+                                <td>{{ $loop->iteration }}</td>
+                                <td> {{ $data->nama }}</td>
+                                <td>
+                                    <a href="eskul/edit/{{ $data->id }}" class="btn btn-warning text-white"><i class="mdi mdi-pencil"></i> Edit</a>
+                                    <a href="eskul/hapus/{{ $data->id }}" class="btn btn-danger text-white" onclick="return confirm('Apakah anda yakin ingin menghapus data ini?')"><i class="mdi mdi-delete"></i> Hapus</a>
+                                </td>
+                            </tr>
+                        @endforeach
+                        </tbody>
+                    </table>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+@endsection
